+++ src/test/php/text/csv/unittest/CsvBeanReaderTest.class.php
@@ -10,12 +10,12 @@ class CsvBeanReaderTest extends CsvReaderTest {
   /**
    * Creates a new CSV reader fixture
    *
-   * @param  io.streams.InputStream $stream
+   * @param  io.streams.Reader|io.streams.InputStream|io.Channel|string $in
    * @param  text.csv.CsvFormat $format
    * @return text.csv.CsvBeanReader
    */
-  protected function newFixture($stream, $format= null) {
-    return new CsvBeanReader(new TextReader($stream), new XPClass(Person::class), $format);
+  protected function newFixture($in, $format= null) {
+    return new CsvBeanReader($in, new XPClass(Person::class), $format);
   }
 
   #[Test]