+++ src/test/php/text/csv/unittest/CsvMapReaderTest.class.php
@@ -9,12 +9,12 @@ class CsvMapReaderTest extends CsvReaderTest {
   /**
    * Creates a new CSV reader fixture
    *
-   * @param  io.streams.InputStream $stream
+   * @param  io.streams.Reader|io.streams.InputStream|io.Channel|string $in
    * @param  text.csv.CsvFormat $format
    * @return text.csv.CsvReader
    */
-  protected function newFixture($stream, $format= null) {
-    return new CsvMapReader(new TextReader($stream), [], $format);
+  protected function newFixture($in, $format= null) {
+    return new CsvMapReader($in, [], $format);
   }
 
   #[Test]